fincoach/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.18.1
requests==2.31.0
python-dotenv==1.0.0
openai-whisper==20231117
pydub==0.25.1
pyaudio==0.2.14
edge-tts==6.1.9
streamlit==1.28.0
